refactored mintBountyModal
